+++ 71_Poker/csharp/Card.cs
@@ -1,0 +1,11 @@
+namespace Poker;
+
+internal record struct Card (Rank Rank, Suit Suit)
+{
+    public override string ToString() => $"{Rank} of {Suit}";
+
+    public static bool operator <(Card x, Card y) => x.Rank < y.Rank;
+    public static bool operator >(Card x, Card y) => x.Rank > y.Rank;
+
+    public static int operator -(Card x, Card y) => x.Rank - y.Rank;
+}
