+++ 71_Poker/csharp/Suit.cs
@@ -1,0 +1,9 @@
+namespace Poker;
+
+internal enum Suit
+{
+    Clubs,
+    Diamonds,
+    Hearts,
+    Spades
+}
